feat:  adjust topic_value to be a double -> this now supports better precision for length and area values

+++ src/main/kotlin/org/heigit/ohsome/now/statsservice/topic/TopicModel.kt
@@ -4,18 +4,18 @@ package org.heigit.ohsome.now.statsservice.topic
 fun Map<String, Any>.toTopicResult(topic: String) = TopicResult(
     this["hashtag"].toString(),
     topic,
-    (this["topic_result"].toString()).toLong()
+    (this["topic_result"].toString()).toDouble()
 )
 
 data class TopicResult(
     val hashtag: String,
     val topic: String,
-    val value: Long,
+    val value: Double,
 )
 
 //TODO: common superclass for all interval results?
 open class TopicIntervalResult(
-    open val value: Long,
+    open val value: Double,
     open val topic: String,
     open val startDate: String,
     open val endDate: String
@@ -26,7 +26,7 @@ fun List<Map<String, Any>>.toTopicIntervalResult(topic: String) = this.map { top
 
 
 fun topicIntervalResult(data: Map<String, Any>, topic: String) = TopicIntervalResult(
-    data["topic_result"].toString().toLong(),
+    data["topic_result"].toString().toDouble(),
     topic,
     data["startdate"].toString(),
     data["enddate"].toString(),
@@ -35,7 +35,7 @@ fun topicIntervalResult(data: Map<String, Any>, topic: String) = TopicIntervalRe
 
 //TODO: common superclass for all interval results?
 open class TopicCountryResult(
-    open val value: Long,
+    open val value: Double,
     open val topic: String,
     open val country: String
 )
@@ -45,20 +45,20 @@ fun List<Map<String, Any>>.toTopicCountryResult(topic: String) = this.map { topi
 
 
 fun topicCountryResult(data: Map<String, Any>, topic: String) = TopicCountryResult(
-    data["topic_result"].toString().toLong(),
+    data["topic_result"].toString().toDouble(),
     topic,
     data["country"].toString()
 )
 
 fun Map<String, Any>.toUserTopicResult(topic: String) = UserTopicResult(
     topic,
-    this["topic_result"].toString().toLong(),
+    this["topic_result"].toString().toDouble(),
     this["user_id"] as Int
 )
 
 
 data class UserTopicResult(
     val topic: String,
-    val value: Long,
+    val value: Double,
     val userId: Int
 )
